refactor(react-components): src/ui を src/primitive にリネーム
ADR-024（プリミティブ層ポリシー）に基づき、最下層を指す語として
`primitive` を採用する。`ui` はあらゆる階層に当てはまる語であり、
最下層の名前として機能していなかった。

- `src/ui/` → `src/primitive/`（git mv で履歴保持）
- `package.json` exports: `./ui` → `./primitive`
- `components.json` の registries パスも更新
- `apps/web` 配下の import パスを一括置換
- 内部相対 import（`src/demo/code-block.tsx`）と
  `link.tsx` の Module Augmentation コメントも更新
