+++ apps/web/src/app/auth/sign-in/_components/auth-error-alert.tsx
@@ -1,4 +1,4 @@
-import { ErrorAlert } from "@next-lift/react-components/ui";
+import { ErrorAlert } from "@next-lift/react-components/primitive";
 import type { FC } from "react";
 
 type Props = {
