+++ apps/web/src/app/cache-demo/05-private-cache/_components/username-form.tsx
@@ -7,7 +7,7 @@ import {
 	DemoCardHeader,
 	DemoCardTitle,
 } from "@next-lift/react-components/demo";
-import { Button } from "@next-lift/react-components/ui";
+import { Button } from "@next-lift/react-components/primitive";
 import type { FC } from "react";
 import { useActionState } from "react";
 import { clearUsername } from "../_mutations/clear-username";
