+++ apps/web/src/app/cache-demo/09-route-handlers/_components/refresh-button.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Button } from "@next-lift/react-components/ui";
+import { Button } from "@next-lift/react-components/primitive";
 import type { FC } from "react";
 import { useActionState } from "react";
 import { revalidateCachedData } from "../_mutations/revalidate-cached-data";
