+++ packages/react-components/components.json
@@ -11,7 +11,7 @@
 	},
 	"aliases": {
 		"components": "@next-lift/react-components/components",
-		"ui": "@next-lift/react-components/ui",
+		"ui": "@next-lift/react-components/primitive",
 		"utils": "@next-lift/react-components/lib/utils",
 		"lib": "@next-lift/react-components/lib",
 		"hooks": "@next-lift/react-components/hooks"
